Refactor modal/dialog components
