+++ kits/Livewire/Base/tests/Feature/Settings/ProfileUpdateTest.php
@@ -60,7 +60,7 @@ class ProfileUpdateTest extends TestCase
 
         $this->actingAs($user);
 
-        $response = Livewire::test('pages::settings.delete-user-form')
+        $response = Livewire::test('pages::settings.delete-user-modal')
             ->set('password', 'password')
             ->call('deleteUser');
 
@@ -78,7 +78,7 @@ class ProfileUpdateTest extends TestCase
 
         $this->actingAs($user);
 
-        $response = Livewire::test('pages::settings.delete-user-form')
+        $response = Livewire::test('pages::settings.delete-user-modal')
             ->set('password', 'wrong-password')
             ->call('deleteUser');
