+++ kits/Livewire/WorkOS/resources/views/pages/settings/delete-user-modal.blade.php
@@ -1,0 +1,39 @@
+<?php
+
+use App\Models\User;
+use Laravel\WorkOS\Http\Requests\AuthKitAccountDeletionRequest;
+use Livewire\Component;
+
+new class extends Component {
+    /**
+     * Delete the currently authenticated user.
+     */
+    public function deleteUser(AuthKitAccountDeletionRequest $request): void
+    {
+        $request->delete(
+            using: fn (User $user) => $user->delete()
+        );
+
+        $this->redirect('/', navigate: true);
+    }
+}; ?>
+
+<flux:modal name="confirm-user-deletion" :show="$errors->isNotEmpty()" focusable class="max-w-lg">
+    <form wire:submit="deleteUser" class="space-y-6">
+        <div>
+            <flux:heading size="lg">{{ __('Are you sure you want to delete your account?') }}</flux:heading>
+
+            <flux:subheading>
+                {{ __('Once your account is deleted, all of its resources and data will also be permanently deleted. Please confirm you would like to permanently delete your account.') }}
+            </flux:subheading>
+        </div>
+
+        <div class="flex justify-end space-x-2 rtl:space-x-reverse">
+            <flux:modal.close>
+                <flux:button variant="filled">{{ __('Cancel') }}</flux:button>
+            </flux:modal.close>
+
+            <flux:button variant="danger" type="submit">{{ __('Delete account') }}</flux:button>
+        </div>
+    </form>
+</flux:modal>
