+++ kits/Livewire/WorkOS/resources/views/pages/settings/profile.blade.php
@@ -1,9 +1,6 @@
 <?php
 
-use App\Models\User;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Session;
-use Illuminate\Validation\Rule;
 use Livewire\Attributes\Title;
 use Livewire\Component;
